refactor(probe): complete dns migration with services and cli hooks
